GP2-901: Move custom Wagtail modeladmin to separate app
This is to partly help prevent core/wagtail_hooks.py getting too big, but we'll also be needign a custom view for the preview of the Case Study anyway, so having that in a CMS-specific app makes sense, too

Changes also disable the tags-related list filter for the CaseStudy modeladmin and makes those tags searchable instead
