+++ cms_extras/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class CmsExtrasConfig(AppConfig):
+    name = 'cms_extras'
+
+    def ready(self):
+        from cms_extras import modeladmin  # noqa F401
